move `free_regions_map` into `infer::outlives`
