+++ src/librustc/infer/outlives/mod.rs
@@ -11,5 +11,6 @@
 //! Various code related to computing outlives relations.
 
 pub mod env;
+pub mod free_region_map;
 pub mod implied_bounds;
 mod obligations;
